+++ crates/optimism/chainspec/src/superchain/mod.rs
@@ -1,0 +1,10 @@
+//! Support for superchain registry.
+
+mod available_chains;
+mod chain_metadata;
+mod chain_spec_macro;
+mod chain_specs;
+mod configs;
+
+pub use available_chains::{AvailableSuperchain, AVAILABLE_CHAINS};
+pub use chain_specs::*;
